Robustness Fixes (#8)
This patch adds infrastructure for integration
like tests. There's a README.md in `bugfixes`
which explains it. In terms of concrete fixes this
patch only addresses bugs 2 and 3 and in general
the graph drawing is much less panicky than
before. In order to achieve that better debugging
was added the to `helper_test.go` which replays
ansi onto a string buffer.

Bug fix 2 and 3 are the same vein through out lots
of the drawing code we now ensure a min/max of the
terminal size to make sure we don't draw things
out of bounds!