Fix Help | Improve Performance (#9)
This patch is another batch of general
improvements. The fun ones to me are the free wins
in performance, by omitting the `.String()` calls
in the `check` calls and leaving that to be lazy
under printf means that we save a ton of resources
not making strings that we just discard.

The general help also got some reading
improvements to closer match the readme
documentation.

* Add godoc comments to `bytes.go`, `check.go` and
  `sliceutils.go`.
* Remove `stubs.go` with a closure we can instead
  not have to hold onto the file handle in this
  stub struct and can instead just delete it
  simplifying it all :)
* Move more things into the draw window this has
  minor perf improvements but moves some of the
  paint order complexity to the draw window.