Themes (#14)
BIG Patch:

# Themes!

This patch codifies the existing theme as `dark`
and comes with 2 more themes which are easily
swappable via `--theme light` and `--theme
complex` finally there's now an option I'd been
meaning to do for a while which is `--theme no`
which puts the program in no colour mode.

Since I was here it was easier to add full support
for custom themes which is covered in separate
docs.

Misc:

* `graph/terminal => terminal` In hindsight this
  should just be a top level package and has
  nothing to do with the graph
* Better logs, a few new messages and beautify a
  few things.
* Confirm that windows and mac work, at least from
  my testing.