Improved Testing (#15)
## Added Rapid Check Dependency: [pgregory.net/rapid](https://github.com/flyingmutant/rapid)

This seemed like a reasonable package and is a new
test dep, in this commit I target the two most
obvious places which I wanted more coverage:
* Serialisation
* Numeric

This has made me about twice as confident that I
don't have bugs in the serialisation, which is
reassuring. However the numeric tests were much
harder to make and don't provide me with much more
confidence.

## Ping API re-touches

This started as just a re-factor but during my
re-factor I noticed some glaring issues - the
client doesn't properly verify sequences and IDs
(still a TODO). The DNS cache only cached a single
IP even if multiple would have worked. I've also
noticed that the listener config stuff is broken
subtly (still a TODO).

The actual code has been shuffled, moving it to a
separate file to make the public API shorter and
clearer.

## GUI Tests

Before this commit the GUI had very limited
testing, this commit changes that to ensure that
`gui.Box` is tested fairly well. We generate some
4000 different input permutations and I hand
checked a few of them for sanity, this helped
catch the missing cases in the positioning code as
well as the missing `heightFromStyle` function. As
explained in-line the actual test-golden's are not
stored in the repo instead we store a hash of all
the test files which is used as short hand.

## Serialisation re-touches

The implementation hasn't changed at all, it has
merely moved to a file per type which makes
navigating the code much nicer, plus lots of
comments.

------

Finally there's some general house-keeping,
version bumps and .gitignore changes.