New Speed up and Slow Down UI (#22)
This patch adds the ability to slow and speed up
the Ping capturing while the program is running.
Before this patch there was no way for the speed
to change while it was running and could only be
configured once via commandline at start up.

When the speed is changing a small up and down
arrow is shown:

https://github.com/user-attachments/assets/05bc6bdc-ffc1-42a4-a9be-031ee1eafc59

### Misc:

* Re-factored GUI to actually be in the right
  package and not in
  `cmd/subcommands/acci-ping/application.go`
* Re-factored Error Toasts to be a shared
  mechanism implemented by the GUI package.
* Fix README and `-l` to be `-debug-log` for
  clarity
* All `-debug*` options are hidden by default
  unless requested by `-help-debug`
* Tweak a few logging calls to be `Info`.
* Fix `computeFrameConfig` evaluation for draw
  calls.
* Fix a bug with the DNS implementation (not
  resetting index)